[CodeClean] separate data utils
Code clean, prepare util functions.
1. separate edge-data utils.
2. remove todo item and extra code clean.

TODO: separate mqtt, edge-event and other util functions.
