+++ tests/unittest_nnstreamer-edge.cc
@@ -9,6 +9,7 @@
 
 #include <gtest/gtest.h>
 #include "nnstreamer-edge.h"
+#include "nnstreamer-edge-data.h"
 #include "nnstreamer-edge-common.h"
 #include "nnstreamer-edge-internal.h"
 #include "nnstreamer-edge-log.h"
